pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shopify-app"
dynamic = ["version"]
description = "Package for building Shopify applications. Authored and maintained by Shopify."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
keywords = ["shopify", "shopify-apps", "token-exchange", "client-credentials", "admin-graphql", "webhooks", "app-proxy", "flow-action"]
authors = [
    { name = "Shopify" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyjwt>=2.8.0",
    "httpx>=0.24.0",
]

[project.urls]
homepage = "https://github.com/Shopify/shopify-app-python/"
issues = "https://community.shopify.dev/c/shopify-cli-libraries/14"
changelog = "https://github.com/Shopify/shopify-app-python/blob/main/CHANGELOG.md"

[tool.hatch.version]
path = "shopify_app/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["shopify_app"]

[tool.hatch.build.targets.sdist]
include = [
    "/shopify_app",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

